+++ R/g_boxplot.R
@@ -41,6 +41,8 @@
 #'   The data inside of the ggplot2 object must also contain the columns with these variable names
 #' @param hline_vars_colors ('character vector') colors for the horizontal lines defined by variables.
 #' @param hline_vars_labels ('character vector') labels for the legend to the horizontal lines defined by variables.
+#' @param replace_y_axis ('logical scalar') flag to determine if y axis break points should be replaced by horizontal
+#'   line values. if FALSE, then new break points will be added instead.
 #'
 #' @author Balazs Toth
 #' @author Jeff Tomlinson (tomlinsj) jeffrey.tomlinson@roche.com
@@ -120,7 +122,8 @@
 #'   hline_arb_label = "Hori_line_label",
 #'   hline_vars = c("ANRHI", "ANRLO", "ULOQN", "LLOQN"),
 #'   hline_vars_colors = c("pink", "brown", "purple", "gray"),
-#'   hline_vars_labels = c("A", "B", "C", "D")
+#'   hline_vars_labels = c("A", "B", "C", "D"),
+#'   replace_y_axis = TRUE
 #' )
 g_boxplot <- function(data,
                       biomarker,
@@ -146,7 +149,9 @@ g_boxplot <- function(data,
                       hline_arb_label = "Horizontal line",
                       hline_vars = character(0),
                       hline_vars_colors = "green",
-                      hline_vars_labels = hline_vars) {
+                      hline_vars_labels = hline_vars,
+                      replace_y_axis = FALSE
+                      ) {
   if (is.null(data[[param_var]])) {
     stop(paste("param_var", param_var, "is not in data."))
   }
@@ -303,5 +308,5 @@ g_boxplot <- function(data,
     hline_arb = hline_arb, hline_arb_color = hline_arb_color, hline_arb_label = hline_arb_label,
     hline_vars = hline_vars, hline_vars_colors = hline_vars_colors, hline_vars_labels = hline_vars_labels
   )
-  line_axis_label(plot1)
+  line_axis_label(plot1, replace_y = replace_y_axis)
 }