+++ R/g_spaghettiplot.R
@@ -35,6 +35,8 @@
 #'   The data inside of the ggplot2 object must also contain the columns with these variable names
 #' @param hline_vars_colors ('character vector') colors for the horizontal lines defined by variables.
 #' @param hline_vars_labels ('character vector') labels for the legend to the horizontal lines defined by variables.
+#' @param replace_y_axis ('logical scalar') flag to determine if y axis break points should be replaced by horizontal
+#'   line values. if FALSE, then new break points will be added instead.
 #'
 #'
 #' @author Wenyi Liu (wenyi.liu@roche.com)
@@ -145,7 +147,8 @@
 #'   group_stats = "median",
 #'   hline_arb = 1.3,
 #'   hline_vars = c("ANRHI", "ANRLO", "ULOQN", "LLOQN"),
-#'   hline_vars_colors = c("pink", "brown", "purple", "gray")
+#'   hline_vars_colors = c("pink", "brown", "purple", "gray"),
+#'   replace_y_axis = TRUE
 #' )
 #'
 #' g_spaghettiplot(
@@ -194,7 +197,8 @@ g_spaghettiplot <- function(data,
                             hline_arb_label = "Horizontal line",
                             hline_vars = character(0),
                             hline_vars_colors = "green",
-                            hline_vars_labels = hline_vars) {
+                            hline_vars_labels = hline_vars,
+                            replace_y_axis = FALSE) {
   checkmate::assert_numeric(ylim, len = 2)
 
   ## Pre-process data
@@ -325,5 +329,5 @@ g_spaghettiplot <- function(data,
     hline_vars_colors = hline_vars_colors,
     hline_vars_labels = hline_vars_labels
   )
-  line_axis_label(plot)
+  line_axis_label(plot, replace_y = replace_y_axis)
 }